add support for the new soft_exceptions field
A new version of logjam_agent will be/is able to send this field. The
semantics of it correspond to a split between the log-levels of the
different exceptions. Any exception can be logged at different
log-levels, usually ranging from info to warning to error. Hard
exceptions (as in the already existing exceptions field) correspond to
errors that were either unrecoverable or that warranted a skip, while
soft-exceptions (usually logged as info or warning) correspond to
exceptions that the code could deal with, e.g. through retries or other
means.