prometheus exporter: use instance labels to fix time series monotonicity problems
In an attempt to reduce label cardinality, the exporter previously used
empty instance labels on its metrics and employed an algorithm to remove
time series that have been inactive for some time. Unfortunately, this
leads to all sorts of problems in Prometheus, because time series can
have gaps which Prometheus can not handle. 

To solve this problem, each collector for a stream maintains a UUID which
is used as an instance label on all metrics for the stream. The UUID gets
regenerated when a given time series gets garbage collected, as described
above, and when a registry needs to be recreated (for example when bucket
sizes change).

This approach also makes sure that a restart of the exporter does not
create discontinuities in time series: every restart will simply create
completely new time series (since the instance labels will be different).