Align logjam spec (#35)
aligned spec  with implementation

* use github ABNF highlighting.

* represent document decimal numbers as decimal

While this currently does not make a difference as
%d1 == %x1 it might reduce confusion.

* visually separate count from OCTET

For fonts that do not clearly separate O from 0
one could mistake it for 40 CTET instead of 4 OCTET.
While the ABNF spec is clear on this topic, additional
readability cannot hurt.

* document tag as byte-sequence (multi-octet).

* allow dashes and underscores in topic names.

* add reference to ABNF documentation.

* add information about bit- & byte-order.

* clarify multi-byte order.

* add ping-pong requirement for shutdown.

* document type for compression-method & version.

* expect at least one segment after topic type

The exception here would be mobile metrics.

* change device-number to 4 OCTETS in consumer

It seems the existing consumers (e.g. logjam-prometheus-exporter)
already parse a 32-bit unsigned integer, which would mean that the bytes
are there.

See also https://github.com/skaes/logjam-tools/blob/master/src/logjam-util.h#L85-L99

* adjust topic spec to make segments optional again

In the future we are likely to migrate to a setup
that only sends "categories" as topics, e.g. `logs` and `events`.